+++ test/testdata/with_pydantic.py
@@ -1,0 +1,14 @@
+"""
+A small example with Pydantic entities.
+"""
+
+import pydantic
+
+
+class Foo(pydantic.BaseModel):
+    """Foo class documentation."""
+
+    a: int = pydantic.Field(default=1, description="Docstring for a")
+
+    b: int = 2
+    """Docstring for b."""